unify data types

+++ sensor_nodes/dht_sensor.h
@@ -8,15 +8,15 @@ bool validUnit(unsigned int i);//needs to get modified if the number of supporte
 int dht_sensor_init(void);
 int dht_cmd(int argc, char **argv);
 
-int sensor_get_temperature(int *temp_i);
+int sensor_get_temperature(int16_t *temp_i);
 Unit sensor_get_temperature_unit(void);
 int* sensor_get_temperature_range(void);
 int sensor_get_temperature_step(void);
 int sensor_get_temperature_precision(void);
 void sensor_set_temperature_unit(Unit u);
-void convertTo(int *temp_i);
+void convertTo(int16_t *temp_i);
 
-int sensor_get_humidity(int *hum_i);
+int sensor_get_humidity(int16_t *hum_i);
 char* sensor_get_humidity_unit(void);
 int sensor_get_humidity_min_range(void);
 int sensor_get_humidity_max_range(void);
